requirements.txt
agate==1.9.1
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.22.10
dbt-common==1.37.3
dbt-core==1.11.8
dbt-duckdb==1.10.1
dbt-extractor==0.6.0
dbt-protos==1.0.443
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
duckdb==1.5.2
Faker==40.15.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==8.9.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
orderly-set==5.5.0
packaging==26.1
pandas==3.0.2
parsedatetime==2.6
pathspec==0.12.1
pillow==12.2.0
plotly==6.7.0
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.13.2
pydantic_core==2.46.2
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
snowplow-tracker==1.1.0
sqlparse==0.5.5
streamlit==1.56.0
tenacity==9.1.4
text-unidecode==1.3
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
zipp==3.23.1
